{"schema":"decodex/commit/1","summary":"Split quantitative per-query row assembly","authority":"manual"}
